Bug fix for decision making during imitation
Changed how the relative ranking is performed for the social network to enforce stronger boundaries. In addition, changed the activity selection to select a subset based on actual ranking and not choose a specific size. So we could get a set of three items for imitation because those are the best three options.